requirements.txt
streamlit
langchain_google_genai
langsmith
python-dotenv